+++ apps/web/src/domain/challenge/daily-mission/DailyMissionSubmitButton.tsx
