+++ apps/web/src/domain/challenge/my-challenge/section/DailyMissionInfoSection.tsx
@@ -1,7 +1,7 @@
 import HybridLink from '@/common/HybridLink';
 import { UserChallengeMissionDetail } from '@/schema';
 import { BONUS_MISSION_TH } from '@/utils/constants';
-import OtVideo from '../../OtVideo';
+import OtVideo from '../../ot-mission/OtVideo';
 import ContentsDropdown from '../dropdown/ContentsDropdown';
 
 interface Props {
