+++ apps/web/src/domain/challenge/my-challenge/section/MissionSubmitRegularSection.tsx
@@ -9,9 +9,9 @@ import { useMissionStore } from '@/store/useMissionStore';
 import { BONUS_MISSION_TH } from '@/utils/constants';
 import { useParams } from 'next/navigation';
 import { useEffect, useState } from 'react';
-import BonusMissionModal from '../../BonusMissionModal';
+import BonusMissionModal from '../../bonus-mission/BonusMissionModal';
 import DashboardCreateReviewModal from '../../dashboard/modal/DashboardCreateReviewModal';
-import MobileReviewModal from '../../MobileReviewModal';
+import MobileReviewModal from '../../modal/MobileReviewModal';
 import MissionSubmitButton from '../mission/MissionSubmitButton';
 import MissionToast from '../mission/MissionToast';
 import LinkInputSection from './LinkInputSection';
