+++ apps/web/src/domain/challenge/old/OldScoreSection.tsx
@@ -2,7 +2,7 @@
 
 import { useRef, useState } from 'react';
 import { useReactToPrint } from 'react-to-print';
-import ScoreTooltipQuestion from '../ScoreTooltipQuestion';
+import ScoreTooltipQuestion from './ScoreTooltipQuestion';
 import CertificatePaper from '../dashboard/CertificatePaper';
 
 interface Props {
