+++ apps/web/src/domain/challenge/old/ScoreTooltipQuestion.tsx
