+++ apps/web/src/domain/challenge/ot-mission/OtMissionInputSection.tsx
