+++ apps/web/src/domain/challenge/ui/NoticeGuideLink.tsx
